requirements.txt
redis
django
pillow
mysqlclient
djangorestframework-jwt
djangorestframework
markdown
django-filter
django-crispy-forms
django-reversion
django-formtools
future
httplib2
six
requests
coreapi
django-cors-headers
xlwt
xlsxwriter
drf-extensions
django-redis
pycryptodome
raven
pygments
social-auth-app-django
pymongo
pykafka
celery